# Avoid Python op creation in commutative cancellation