--- a/crates/circuit/src/circuit_instruction.rs
+++ b/crates/circuit/src/circuit_instruction.rs
@@ -15,7 +15,7 @@ use std::cell::RefCell;
 
 use numpy::IntoPyArray;
 use pyo3::basic::CompareOp;
-use pyo3::exceptions::{PyDeprecationWarning, PyTypeError, PyValueError};
+use pyo3::exceptions::{PyDeprecationWarning, PyValueError};
 use pyo3::prelude::*;
 use pyo3::types::{IntoPyDict, PyList, PyTuple, PyType};
 use pyo3::{intern, IntoPy, PyObject, PyResult};
@@ -430,41 +430,6 @@ impl CircuitInstruction {
     fn matrix(&self, py: Python) -> Option<PyObject> {
         let matrix = self.operation.matrix(&self.params);
         matrix.map(|mat| mat.into_pyarray_bound(py).into())
-    }
-
-    fn to_matrix(&self, py: Python) -> Option<PyObject> {
-        self.matrix(py)
-    }
-
-    fn __array__(
-        &self,
-        py: Python,
-        dtype: Option<PyObject>,
-        copy: Option<bool>,
-    ) -> PyResult<PyObject> {
-        if copy == Some(false) {
-            return Err(PyValueError::new_err(
-                "A copy is needed to return an array from this object.",
-            ));
-        }
-        let res =
-            match self.to_matrix(py) {
-                Some(res) => res,
-                None => return Err(PyTypeError::new_err(
-                    "ParameterExpression with unbound parameters cannot be represented as an array",
-                )),
-            };
-        Ok(match dtype {
-            Some(dtype) => {
-                let numpy_mod = py.import_bound("numpy")?;
-                let args = (res,);
-                let kwargs = [("dtype", dtype)].into_py_dict_bound(py);
-                numpy_mod
-                    .call_method("asarray", args, Some(&kwargs))?
-                    .into()
-            }
-            None => res,
-        })
     }
 
     #[getter]