--- a/qiskit/quantum_info/operators/operator.py
+++ b/qiskit/quantum_info/operators/operator.py
@@ -112,6 +112,7 @@ class Operator(LinearOp):
             # a complex numpy matrix.
             self._data = np.asarray(data.to_matrix(), dtype=complex)
         else:
+            print(data)
             raise QiskitError("Invalid input data format for Operator")
 
         super().__init__(